chore(): move `dir` directive to core
