+++ src/core/rtl/dir.ts
@@ -1,6 +1,6 @@
 import {EventEmitter} from 'angular2/src/facade/async';
 import {Directive, HostBinding, Output, Input} from 'angular2/core';
-import {OneOf} from '../../core/annotations/one-of';
+import {OneOf} from '../annotations/one-of';
 
 
 /**
