+++ src/index.html
@@ -30,10 +30,6 @@
           format: 'register',
           defaultExtension: 'js'
         },
-        'directives': {
-          format: 'register',
-          defaultExtension: 'js'
-        },
         'core': {
           format: 'register',
           defaultExtension: 'js'
